% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{retrieve_ento_data}
\alias{retrieve_ento_data}
\title{helpers}
\usage{
retrieve_ento_data(
  s3creds_path = "../../credentials/bohemiacensuss3credentials.csv",
  country = "Mozambique"
)
}
\arguments{
\item{s3creds_path}{The path to a csv file with s3 credentials}

\item{country}{Mozambique or Tanzania}
}
\value{
The return value, if any, from executing the function.

An object loaded into memory
}
\description{
A fct function
}
