% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{write_ento_data}
\alias{write_ento_data}
\title{Write ento data}
\usage{
write_ento_data(
  full_a4,
  directory,
  s3creds_path = "../../credentials/bohemiacensuss3credentials.csv",
  country = "Mozambique"
)
}
\arguments{
\item{s3creds_path}{The path to a csv file with s3 credentials}

\item{country}{Mozambique or Tanzania}

\item{a4_submission}{A three column dataframe}
}
\value{
An object loaded into memory
}
\description{
Write entomology data to AWS s3
}
